+++ weldx/core/__init__.py
@@ -1,1 +1,2 @@
 from .math_expression import MathematicalExpression
+from .spatial_series import SpatialSeries
